auto detail adjustments to Space, instagram, favicon shrink

+++ lib/Space.vert
@@ -1,4 +1,4 @@
-precision highp float;
+precision lowp float;
 attribute vec2 a_position;
 void main() {
     gl_Position = vec4(a_position, 0, 1);
